Flicky thing fixed and more!
Like it says I finally fixed the flicky thing, the problem was that the structures weren't able to load within the amount of time it takes to register that it has indeed been generated in that spot so it would stick to that one spot. I also (hopefully haven't tested yet) have fixed the glitch when entering the dimension where it doesn't load anything in multiplayer due to lag.